수정 사항
